requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-experimental>=0.0.50
psycopg2-binary>=2.9.7
sqlalchemy>=2.0.23
python-dotenv>=1.0.0
pandas>=2.1.0
plotly>=5.17.0
numpy>=1.24.0 